requirements.txt
# YellowOracle dependencies
# Sync script
supabase>=2.27.0
requests>=2.32.0
python-dotenv>=1.0.0
aiohttp>=3.9.0  # Per parallelizzazione API

# MCP server
mcp>=1.26.0

# Dashboard (optional, not needed for sync)
# streamlit>=1.53.0
# altair>=6.0.0
